+++ src/scrollview/HISTORY.md
@@ -3,6 +3,8 @@ ScrollView Change History
 
 3.7.0
 -----
+  * Moved: _isOOB to _isOutOfBounds
+  
   * Moved: SNAP_DURATION static property (deprecated) to 'snapDuration' ATTR
 
   * Moved: SNAP_EASING static property (deprecated) to 'snapEasing' ATTR